requirements.txt
IMDbPY
openai
prettytable
wcwidth
python-dotenv